refactor: 파일 구조 변경
from:
domain

to:
layered
