+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -1,0 +1,15 @@
+//package roomescape.controller;
+//
+//import org.springframework.web.bind.annotation.GetMapping;
+//import org.springframework.web.bind.annotation.RequestMapping;
+//import org.springframework.web.bind.annotation.RestController;
+//
+//@RequestMapping("/admin/reservations")
+//@RestController
+//public class AdminReservationController {
+//    private final AdminReservationService adminReservationService;
+//
+//    @GetMapping
+//
+//
+//}
