+++ src/main/java/roomescape/domain/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime;
+package roomescape.domain;
 
 import java.time.LocalTime;
