+++ src/main/java/roomescape/repository/ReservationTimeDao.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime;
+package roomescape.repository;
 
 import java.time.LocalTime;
 import java.util.List;
@@ -8,6 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
 
 @Repository
 public class ReservationTimeDao {
