+++ src/test/java/roomescape/reservationTime/ReservationTimeTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalTime;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.ReservationTime;
 
 class ReservationTimeTest {
