+++ src/test/java/roomescape/reservationTime/dto/ReservationTimeResponseTest.java
@@ -4,7 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalTime;
 import org.junit.jupiter.api.Test;
-import roomescape.reservationtime.ReservationTime;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeResponse;
 
 class ReservationTimeResponseTest {
